Add JsonFormsController and RecordController.
This resolves a lot of implied methods within the Controller class.

A frontend controller can now fully use action logs and json forms with full type safety.
